--- integrations/opensearch/src/haystack_integrations/common/opensearch/errors.py
@@ -1,5 +1,0 @@
-from haystack.document_stores.errors import DocumentStoreError
-
-
-class AWSConfigurationError(DocumentStoreError):
-    """Exception raised when AWS is not configured correctly"""
