+++ integrations/opensearch/src/haystack_integrations/document_stores/opensearch/auth.py
@@ -2,10 +2,64 @@ from dataclasses import dataclass, field, fields
 from typing import Any, Dict, Optional
 
 from haystack import default_from_dict, default_to_dict
+from haystack.document_stores.errors import DocumentStoreError
+from haystack.lazy_imports import LazyImport
 from haystack.utils.auth import Secret, deserialize_secrets_inplace
-from haystack_integrations.common.opensearch.errors import AWSConfigurationError
-from haystack_integrations.common.opensearch.utils import get_aws_session
 from opensearchpy import Urllib3AWSV4SignerAuth
+
+with LazyImport("Run 'pip install \"boto3\"' to install boto3.") as boto3_import:
+    import boto3
+    from botocore.exceptions import BotoCoreError
+
+
+AWS_CONFIGURATION_KEYS = [
+    "aws_access_key_id",
+    "aws_secret_access_key",
+    "aws_session_token",
+    "aws_region_name",
+    "aws_profile_name",
+]
+
+
+class AWSConfigurationError(DocumentStoreError):
+    """Exception raised when AWS is not configured correctly"""
+
+
+def _get_aws_session(
+    aws_access_key_id: Optional[str] = None,
+    aws_secret_access_key: Optional[str] = None,
+    aws_session_token: Optional[str] = None,
+    aws_region_name: Optional[str] = None,
+    aws_profile_name: Optional[str] = None,
+    **kwargs,
+):
+    """
+    Creates an AWS Session with the given parameters.
+    Checks if the provided AWS credentials are valid and can be used to connect to AWS.
+
+    :param aws_access_key_id: AWS access key ID.
+    :param aws_secret_access_key: AWS secret access key.
+    :param aws_session_token: AWS session token.
+    :param aws_region_name: AWS region name.
+    :param aws_profile_name: AWS profile name.
+    :param kwargs: The kwargs passed down to the service client. Supported kwargs depend on the model chosen.
+        See https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters.html.
+    :raises AWSConfigurationError: If the provided AWS credentials are invalid.
+    :returns: The created AWS session.
+    """
+    boto3_import.check()
+    try:
+        return boto3.Session(
+            aws_access_key_id=aws_access_key_id,
+            aws_secret_access_key=aws_secret_access_key,
+            aws_session_token=aws_session_token,
+            region_name=aws_region_name,
+            profile_name=aws_profile_name,
+        )
+    except BotoCoreError as e:
+        provided_aws_config = {k: v for k, v in kwargs.items() if k in AWS_CONFIGURATION_KEYS}
+        msg = f"Failed to initialize the session with provided AWS credentials {provided_aws_config}"
+        raise AWSConfigurationError(msg) from e
 
 
 @dataclass()
@@ -83,7 +137,7 @@ class AWSAuth:
 
         try:
             region_name = resolve_secret(self.aws_region_name)
-            session = get_aws_session(
+            session = _get_aws_session(
                 aws_access_key_id=resolve_secret(self.aws_access_key_id),
                 aws_secret_access_key=resolve_secret(self.aws_secret_access_key),
                 aws_session_token=resolve_secret(self.aws_session_token),
